Refactor telemetry boundary cleanup
