+++ controller/include/telemetry/telemetry_snapshot_builder.h
@@ -6,10 +6,15 @@
 
 #include <nlohmann/json.hpp>
 
+#include "telemetry/telemetry_alert_builder.h"
 #include "telemetry/telemetry_frame_matcher.h"
-#include "telemetry/telemetry_health_builder.h"
+#include "telemetry/telemetry_frame_json_builder.h"
+#include "telemetry/telemetry_history_downsampler.h"
 #include "telemetry/telemetry_live_store_types.h"
+#include "telemetry/telemetry_node_health_builder.h"
+#include "telemetry/telemetry_plane_aggregate_builder.h"
 #include "telemetry/telemetry_persistence_repository.h"
+#include "telemetry/telemetry_state_types.h"
 #include "telemetry/telemetry_stream_metrics_service.h"
 
 namespace naim::controller {
@@ -27,22 +32,14 @@ class TelemetrySnapshotBuilder final {
       const std::optional<std::string>& plane_name,
       int history_seconds,
       std::uint64_t now_ms) const;
-  nlohmann::json FrameToJson(
-      const naim::HostTelemetryFrame& frame,
-      std::uint64_t now_ms) const;
-  nlohmann::json BuildLatencyBreakdown(
-      const naim::HostTelemetryFrame& frame,
-      std::uint64_t controller_ingest_delay_ms) const;
-  nlohmann::json BuildDownsampledHistory(
-      const std::vector<const TelemetryNodeBuffer*>& buffers,
-      int history_seconds,
-      std::uint64_t now_ms,
-      std::uint64_t warm_bucket_ms,
-      std::uint64_t cold_bucket_ms) const;
 
  private:
+  TelemetryAlertBuilder alert_builder_;
   TelemetryFrameMatcher matcher_;
-  TelemetryHealthBuilder health_builder_;
+  TelemetryFrameJsonBuilder frame_json_builder_;
+  TelemetryHistoryDownsampler history_downsampler_;
+  TelemetryNodeHealthBuilder node_health_builder_;
+  TelemetryPlaneAggregateBuilder plane_aggregate_builder_;
   TelemetryPersistenceRepository persistence_repository_;
   TelemetryStreamMetricsService stream_metrics_service_;
 };
