+++ controller/src/telemetry/telemetry_live_store.cpp
@@ -1,8 +1,16 @@
 #include "telemetry/telemetry_live_store.h"
 
-#include <algorithm>
+#include <utility>
+
+#include "telemetry/telemetry_live_store_state.h"
+#include "telemetry/telemetry_state_types.h"
 
 namespace naim::controller {
+
+TelemetryLiveStore::TelemetryLiveStore()
+    : state_(std::make_unique<TelemetryLiveStoreState>()) {}
+
+TelemetryLiveStore::~TelemetryLiveStore() = default;
 
 TelemetryLiveStore& TelemetryLiveStore::Instance() {
   static TelemetryLiveStore store;
@@ -18,9 +26,9 @@ void TelemetryLiveStore::ConfigurePersistence(
 
 void TelemetryLiveStore::ConfigureFromEnvironment(const std::string& db_path) {
   const auto retention =
-      operational_config_.RetentionFromEnvironment(retention_);
+      state_->operational_config.RetentionFromEnvironment(state_->retention);
   const auto thresholds =
-      operational_config_.ThresholdsFromEnvironment(thresholds_);
+      state_->operational_config.ThresholdsFromEnvironment(state_->thresholds);
   ConfigureOperationalPolicy(retention, thresholds);
   ConfigurePersistence(db_path, retention.durable_history_capacity);
 }
@@ -29,31 +37,34 @@ void TelemetryLiveStore::ConfigureOperationalPolicy(
     RetentionConfig retention,
     AlertThresholds thresholds) {
   std::lock_guard<std::mutex> lock(mutex_);
-  retention_ = operational_config_.NormalizeRetention(retention);
-  thresholds_ = thresholds;
-  ring_buffer_.ApplyHotRetention(nodes_, dropped_frames_total_, retention_);
+  state_->retention = state_->operational_config.NormalizeRetention(retention);
+  state_->thresholds = thresholds;
+  state_->ring_buffer.ApplyHotRetention(
+      state_->nodes,
+      state_->dropped_frames_total,
+      state_->retention);
 }
 
 void TelemetryLiveStore::ResetForTests() {
   std::lock_guard<std::mutex> lock(mutex_);
-  nodes_.clear();
-  latest_sequence_ = 0;
-  dropped_frames_total_ = 0;
-  retention_ = RetentionConfig{};
-  thresholds_ = AlertThresholds{};
-  persistence_ = TelemetryPersistenceState{};
-  streams_ = TelemetryStreamMetrics{};
+  state_->nodes.clear();
+  state_->latest_sequence = 0;
+  state_->dropped_frames_total = 0;
+  state_->retention = RetentionConfig{};
+  state_->thresholds = AlertThresholds{};
+  state_->persistence = TelemetryPersistenceState{};
+  state_->streams = TelemetryStreamMetrics{};
 }
 
 bool TelemetryLiveStore::Upsert(naim::HostTelemetryFrame frame) {
   if (frame.node_name.empty()) {
     return false;
   }
-  frame = frame_normalizer_.Normalize(std::move(frame));
+  frame = state_->frame_normalizer.Normalize(std::move(frame));
   std::lock_guard<std::mutex> lock(mutex_);
   const bool updated = UpsertInMemoryLocked(frame);
   if (updated) {
-    persistence_repository_.PersistFrame(persistence_, frame);
+    state_->persistence_repository.PersistFrame(state_->persistence, frame);
   }
   return updated;
 }
@@ -62,90 +73,94 @@ nlohmann::json TelemetryLiveStore::BuildSnapshot(
     const std::optional<std::string>& plane_name,
     const int history_seconds) const {
   std::lock_guard<std::mutex> lock(mutex_);
-  return snapshot_builder_.BuildSnapshot(
-      nodes_,
-      retention_,
-      thresholds_,
-      persistence_,
-      streams_,
-      latest_sequence_,
-      dropped_frames_total_,
+  return state_->snapshot_builder.BuildSnapshot(
+      state_->nodes,
+      state_->retention,
+      state_->thresholds,
+      state_->persistence,
+      state_->streams,
+      state_->latest_sequence,
+      state_->dropped_frames_total,
       plane_name,
       history_seconds,
-      clock_.CurrentUnixMillis());
+      state_->clock.CurrentUnixMillis());
 }
 
 nlohmann::json TelemetryLiveStore::BuildHealth(
     const std::optional<std::string>& plane_name) const {
   std::lock_guard<std::mutex> lock(mutex_);
-  return health_builder_.BuildHealth(
-      nodes_,
-      retention_,
-      thresholds_,
-      persistence_,
-      streams_,
-      latest_sequence_,
-      dropped_frames_total_,
+  return state_->health_builder.BuildHealth(
+      state_->nodes,
+      state_->retention,
+      state_->thresholds,
+      state_->persistence,
+      state_->streams,
+      state_->latest_sequence,
+      state_->dropped_frames_total,
       plane_name,
-      clock_.CurrentUnixMillis());
+      state_->clock.CurrentUnixMillis());
 }
 
 std::string TelemetryLiveStore::BuildOpenMetrics(
     const std::optional<std::string>& plane_name) const {
-  return open_metrics_exporter_.Build(BuildHealth(plane_name), plane_name);
+  return state_->open_metrics_exporter.Build(BuildHealth(plane_name), plane_name);
 }
 
 std::vector<naim::HostTelemetryFrame> TelemetryLiveStore::LoadFramesAfter(
     const std::uint64_t sequence,
     const std::optional<std::string>& plane_name) const {
   std::lock_guard<std::mutex> lock(mutex_);
-  return ring_buffer_.LoadFramesAfter(nodes_, retention_, sequence, plane_name);
+  return state_->ring_buffer.LoadFramesAfter(
+      state_->nodes,
+      state_->retention,
+      sequence,
+      plane_name);
 }
 
 TelemetryLiveStore::StreamDelta TelemetryLiveStore::LoadStreamDeltaAfter(
     const std::uint64_t sequence,
     const std::optional<std::string>& plane_name) const {
   std::lock_guard<std::mutex> lock(mutex_);
-  return ring_buffer_.LoadStreamDeltaAfter(
-      nodes_,
-      retention_,
-      latest_sequence_,
-      dropped_frames_total_,
+  return state_->ring_buffer.LoadStreamDeltaAfter(
+      state_->nodes,
+      state_->retention,
+      state_->latest_sequence,
+      state_->dropped_frames_total,
       sequence,
       plane_name);
 }
 
 std::uint64_t TelemetryLiveStore::LatestSequence() const {
   std::lock_guard<std::mutex> lock(mutex_);
-  return latest_sequence_;
+  return state_->latest_sequence;
 }
 
 void TelemetryLiveStore::RecordStreamOpened(const std::string& stream_name) {
   std::lock_guard<std::mutex> lock(mutex_);
-  stream_metrics_service_.RecordOpened(streams_, stream_name);
+  state_->stream_metrics_service.RecordOpened(state_->streams, stream_name);
 }
 
 void TelemetryLiveStore::RecordStreamClosed(const std::string& stream_name) {
   std::lock_guard<std::mutex> lock(mutex_);
-  stream_metrics_service_.RecordClosed(streams_, stream_name);
+  state_->stream_metrics_service.RecordClosed(state_->streams, stream_name);
 }
 
 void TelemetryLiveStore::RecordStreamReplayRequired(const std::string& stream_name) {
   std::lock_guard<std::mutex> lock(mutex_);
-  stream_metrics_service_.RecordReplayRequired(streams_, stream_name);
+  state_->stream_metrics_service.RecordReplayRequired(state_->streams, stream_name);
 }
 
 void TelemetryLiveStore::RecordStreamSendFailure(const std::string& stream_name) {
   std::lock_guard<std::mutex> lock(mutex_);
-  stream_metrics_service_.RecordSendFailure(streams_, stream_name);
+  state_->stream_metrics_service.RecordSendFailure(state_->streams, stream_name);
 }
 
 bool TelemetryLiveStore::UpsertInMemoryLocked(naim::HostTelemetryFrame frame) {
-  return ring_buffer_.Upsert(
-      nodes_,
-      latest_sequence_,
-      dropped_frames_total_,
-      retention_,
+  return state_->ring_buffer.Upsert(
+      state_->nodes,
+      state_->latest_sequence,
+      state_->dropped_frames_total,
+      state_->retention,
       std::move(frame));
 }
 
@@ -153,9 +168,12 @@ void TelemetryLiveStore::ConfigurePersistenceLocked(
     const std::string& db_path,
     const std::size_t retention_capacity) {
   const auto frames =
-      persistence_repository_.Configure(persistence_, db_path, retention_capacity);
+      state_->persistence_repository.Configure(
+          state_->persistence,
+          db_path,
+          retention_capacity);
   for (auto frame : frames) {
-    frame = frame_normalizer_.Normalize(std::move(frame));
+    frame = state_->frame_normalizer.Normalize(std::move(frame));
     UpsertInMemoryLocked(std::move(frame));
   }
 }
