+++ controller/src/telemetry/telemetry_live_store_state.cpp
@@ -1,0 +1,7 @@
+#include "telemetry/telemetry_live_store_state.h"
+
+namespace naim::controller {
+
+TelemetryLiveStoreState::TelemetryLiveStoreState() = default;
+
+}  // namespace naim::controller
